FIX: Webpoda downloads shouldn't be impacted by duplicate uploads
When duplicate uploads occur to the IMAP SDC, we would throw an
IMAPDataAccessError which would propagate up and stop all other
download/uploads for other days of this instrument, or other
instruments when used in a loop. We don't necessarily need to
raise this error, but can rather just log it as an error and move
on attempting all of the other days too.