FIX: Add data versioning to webpoda packet file generation
To account for partial days we need to query to see whether there
are any previously existing L0 files for this specific date/repointing.
We then increment by 1 after that. This accounts for any potential
partial day downlinks where we then get the next half of the day the
following downlink, or backfilled gaps.