FIX: Add end_date to query to restrict future file searches
We only passed in start_date to our query before, but this does
a >= search into the future, which meant that during our I&T testing
when more files are uploaded and in the future, we can accidentally
match on them. We only want the specific day of this file, so add
the end_date parameter as well.